Work In Progress - Begin adding config
src/main/resources/TBM_Config.xml will be copied to this mod's config
directory. If a config file already exists and the version number is the
same, it won't be overridded. If the version number is less than the
mod's internal config file, then the old one will be renamed and the new
one will take its place. At least this is how it is planned.