Impl ignore specific damage sources in battle
Fixes #1. (Damage is inhibited, but some things like amount of air while
underwater still decreases for the duration of being underwater regardless of
being in battle.)

Server-side config specifies which damage sources are ignored in battle. This
can also be set in-game with `/tbm-server-edit`.

Also updated config for entities which should reduce the amount of stuff printed
into the logs (sorry about that).